requirements.txt
streamlit==1.32.0
langchain==0.1.13
langchain-google-genai==1.0.1
chromadb==0.4.24
pypdf==4.1.0
python-dotenv==1.0.1